api/requirements.txt
fastapi
fastapi-jwt-auth
sqlmodel
psycopg2-binary
redis
passlib
python-multipart